Addressed review comments. Moved ProxyPorts to Module interface. Changed location for Module tests.
