+++ engine/src/test/java/com/datatorrent/stram/plan/logical/module/FilterOperator.java
@@ -16,52 +16,29 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.datatorrent.stram.moduleexperiment.testModule;
+package com.datatorrent.stram.plan.logical.module;
 
-import java.util.Random;
-
-import com.datatorrent.api.Context.OperatorContext;
+import com.datatorrent.api.DefaultInputPort;
 import com.datatorrent.api.DefaultOutputPort;
-import com.datatorrent.api.InputOperator;
+import com.datatorrent.common.util.BaseOperator;
 
 /**
- * Toy Random Input Operator.
- * Generates random integers
+ * Toy Filter Operator. Removes negative values
  */
-public class RandomInputOperator implements InputOperator
+public class FilterOperator extends BaseOperator
 {
-
-  Random r;
-  public transient DefaultOutputPort<Integer> output = new DefaultOutputPort<Integer>();
-  long sentAt = System.currentTimeMillis();
-
-  @Override
-  public void beginWindow(long windowId)
-  {
-  }
-
-  @Override
-  public void endWindow()
-  {
-  }
-
-  @Override
-  public void setup(OperatorContext context)
-  {
-    r = new Random();
-  }
-
-  @Override
-  public void teardown()
-  {
-  }
-
-  @Override
-  public void emitTuples()
-  {
-    if(System.currentTimeMillis() - sentAt > 100){
-      output.emit(r.nextInt());
-      sentAt = System.currentTimeMillis();
+  public transient DefaultInputPort<Integer> input = new DefaultInputPort<Integer>() {
+    
+    @Override
+    public void process(Integer tuple)
+    {
+      if(tuple.intValue() >= 0)
+      {
+        output.emit(tuple);
+      }
     }
-  }
+  };
+
+  public transient DefaultOutputPort<Integer> output = new DefaultOutputPort<Integer>();
 }
+
