+++ engine/src/test/java/com/datatorrent/stram/plan/logical/module/InnerModule.java
@@ -16,36 +16,41 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.datatorrent.stram.moduleexperiment.testModule;
+package com.datatorrent.stram.plan.logical.module;
 
 import org.apache.hadoop.conf.Configuration;
 
 import com.datatorrent.api.DAG;
 import com.datatorrent.api.Module;
-import com.datatorrent.api.Operator.ProxyInputPort;
-import com.datatorrent.api.Operator.ProxyOutputPort;
 
 /**
- * Outer module - Level 1
- * Has an Operator - FilterOperator and a Module - OddEvenModule (Level 2)
+ * Inner Module - Level 2
+ * Has an Operator which splits the stream into odd and even integers
  */
-public class OuterModule implements Module
+public class InnerModule implements Module
 {
-
+  /*
+   * Proxy ports for the module
+   * mInput - Input Proxy Port
+   * mOutputEven, mOutputOdd - Output Proxy Ports
+   */
   public transient ProxyInputPort<Integer> mInput = new ProxyInputPort<Integer>();
   public transient ProxyOutputPort<Integer> mOutputEven = new ProxyOutputPort<Integer>();
   public transient ProxyOutputPort<Integer> mOutputOdd = new ProxyOutputPort<Integer>();
 
+  /**
+   * populateDag() method for the module.
+   * Called when expanding the logical Dag
+   */
   @Override
   public void populateDAG(DAG dag, Configuration conf)
   {
-    FilterOperator filter = dag.addOperator("FilterOperator", new FilterOperator());
-    mInput.set(filter.input);
-
-    InnerModule innerModule = dag.addModule("InnerModule", new InnerModule());
-    mOutputEven.set(innerModule.mOutputEven);
-    mOutputOdd.set(innerModule.mOutputOdd);
-
-    dag.addStream("FilterToInnerModule", filter.output, innerModule.mInput);
+    OddEvenOperator moduleOperator = dag.addOperator("InnerModule", new OddEvenOperator());
+    /*
+     * Set the operator ports inside the proxy port objects
+     */
+    mInput.set(moduleOperator.input);
+    mOutputEven.set(moduleOperator.even);
+    mOutputOdd.set(moduleOperator.odd);
   }
 }
